backend/requirements.txt
fastapi
uvicorn
joblib
huggingface_hub
requests
python-dotenv
scikit-learn==1.6.1
